% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateset.R
\name{addition}
\alias{addition}
\title{Using regression to find the network matrix}
\usage{
addition(data, lambda, P)
}
\arguments{
\item{data}{Data matrix}

\item{lambda}{Vector of hyperparameters}

\item{P}{The prior adjacency matrix}
}
\value{
An initial matrix
}
\description{
INTERNAL METHOD: Use sparse regression methods to estimate the adjacency matrix
}
