% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Evaluate real vs estimated network}
\usage{
comparison(real, estimate)
}
\arguments{
\item{real}{The real network}

\item{estimate}{The estimated network using \code{SBIC}}
}
\description{
Evaluate real vs estimated network
}
