% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateset.R
\name{deletion}
\alias{deletion}
\title{Update prior network graph with data and remove edges if necessary}
\usage{
deletion(data, lambda, P)
}
\arguments{
\item{data}{Data to be input}

\item{lambda}{A vector of hyperparameters}

\item{P}{A prior network graph}
}
\description{
Update prior network graph with data and remove edges if necessary
}
