% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{emMain}
\alias{emMain}
\alias{emMain.LatentForests}
\title{EM-algorithm for latent forests.}
\usage{
emMain(this, model, starts, maxIter, tol)

\method{emMain}{LatentForests}(this, model, starts=5, maxIter=1000, tol=1e-4)
}
\arguments{
\item{this}{the LatentForests object.}

\item{model}{the model for which to compute the approximate MLE.}

\item{starts}{the number of random restarts.}

\item{maxIter}{the maximum number of iterations to complete in the algorithm.}

\item{tol}{the tolerance to use a convergence criterion.}
}
\description{
Uses the EM-algorithm (with multiple random restarts) to compute an
approximate maximum likelihood estimate for a given latent forest model.
}

