% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{emSteps}
\alias{emSteps}
\alias{emSteps.LatentForests}
\title{One EM-iteration.}
\usage{
emSteps(this, support, S)

\method{emSteps}{LatentForests}(this, support, S)
}
\arguments{
\item{this}{the LatentForests object.}

\item{support}{the support of the model.}

\item{S}{the current covariance matrix.}
}
\description{
A private method that performs a single iteration of the EM-algorithm, this
is a helper function for emMain method.
}

