% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalyses.R
\name{FactorAnalyses}
\alias{FactorAnalyses}
\title{Construct a poset of factor analysis models.}
\usage{
FactorAnalyses(numCovariates = 1, maxNumFactors = 0)
}
\arguments{
\item{numCovariates}{the number of covariates in all of the models.}

\item{maxNumFactors}{the maximum number of factors allowed in a model, will
create a hierarchy of all models with less than or equal
to this number.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of factor analysis models which
have a natural ordering upon them.
}

