% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtures.R
\name{GaussianMixtures}
\alias{GaussianMixtures}
\title{Construct a poset of gaussian mixture models.}
\usage{
GaussianMixtures(maxNumComponents = 1, learnCoefBound = "")
}
\arguments{
\item{maxNumComponents}{the maximum number of gaussian components to
consider in a mixture.}

\item{learnCoefBound}{which bound on learning coefficients to use, can be
one of
\itemize{
 \item{"1"}
 \item{"+"}
}}

\item{restarts}{the number of random restarts to perform when computing the
MLE.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of gaussian mixture models which
have a natural ordering upon them.
}

