% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbic.R
\name{gbic}
\alias{gbic}
\title{SBIC criterion for Gaussian Graphical Model}
\usage{
gbic(data, theta, tem, P)
}
\arguments{
\item{data}{A data in matrix form of \code{n} observations and \code{p} variables}

\item{theta}{A \code{p} by \code{p} matrix representing the given graph structure}

\item{tem}{A temperature parameter involved in SBIC}

\item{P}{The prior adjacency matrix}
}
\value{
\item{gbic}{the value of gbic with given temperature parameter and prior adjacency matrix}
}
\description{
This function
}
\examples{
 # set.seed(1)
 # d=simulate(p=10,n=100, prob1 = 0.1, prob2 = 0.2,ka=4)
 # data=d$data
 # P=d$priornetwork
 # theta=d$realnetwork
 # tem=0.5
 # index=sbic(data=data, theta=theta, tem=tem, P=P)
}
\author{
Jie Zhou
}
