% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests-Helpers.R
\name{generateAllBinaryTrees}
\alias{generateAllBinaryTrees}
\title{Generate all non-isomorphic binary trees.}
\usage{
generateAllBinaryTrees(numLeaves)
}
\arguments{
\item{numLeaves}{the number of leaves}
}
\description{
Generates all non-isomorphic binary trees with a given number of
leaves where leaves are considered labeled and inner nodes are
unlabeled. Takes as argument the number of leaves for which to produce the
binary trees and returns a list of (n-1)x2 matrices where each row
corresponds to a edge in the tree. These edge matrices will be in 'directed
order,' i.e. will be so that if they are considered to be directed edges
then the resulting graph will have exactly one source.
}

