% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getAllEdges}
\alias{getAllEdges}
\alias{getAllEdges.LatentForests}
\title{Edges representing the largest model.}
\usage{
getAllEdges(this, model)

\method{getAllEdges}{LatentForests}(this, model)
}
\arguments{
\item{this}{the LatentForests object.}

\item{model}{the model number.}
}
\description{
When creating the LatentForests object a set of edges representing the
largest model is required. This function returns those edges as a matrix.
This matrix will have edges in the same order but may have flipped which
node comes first in any particular edge. That is if edge (1,4) was the
5th edge then it will remain the 5th edge but may now be of the form (4,1).
}

