% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getCovMat}
\alias{getCovMat}
\alias{getCovMat.LatentForests}
\title{Create a covariance matrix.}
\usage{
getCovMat(this, edgeCorrelations)

\method{getCovMat}{LatentForests}(this, edgeCorrelations)
}
\arguments{
\item{this}{the LatentForests object.}

\item{edgeCorrelations}{the edge correlations in a numeric vector.}
}
\description{
Creates a covariance matrix for the latent forest model where edge
correlations are given. Here edge correlations are given as a vector and
correspond (in order) to the edges returned by this$getAllEdges().
}

