% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ModelPoset.R, R/ReducedRankRegressions.R
\name{getData.BinomialMixtures}
\alias{getData}
\alias{getData.BinomialMixtures}
\alias{getData.FactorAnalyses}
\alias{getData.GaussianMixtures}
\alias{getData.LCAs}
\alias{getData.LatentForests}
\alias{getData.ReducedRankRegressions}
\title{Return the set data.}
\arguments{
\item{this}{the BinomialMixtures object}
}
\description{
If data has been set for the object using the getData() function
then will return that data, otherwise will throw an error.
}

