% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getModelWithSupport}
\alias{getModelWithSupport}
\alias{getModelWithSupport.LatentForests}
\title{Get model with the given support.}
\usage{
getModelWithSupport(this, support)

\method{getModelWithSupport}{LatentForests}(this, support)
}
\arguments{
\item{this}{the LatentForests object.}

\item{support}{the 0-1 vector representing the support.}
}
\description{
Returns the model number corresponding to a given 0-1 vector representing
the support of the model. This support should corresponds to the edges
returned by this$getAllEdges()
}

