% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalyses.R
\name{getNumFactorsForModel}
\alias{getNumFactorsForModel}
\alias{getNumFactorsForModel.FactorAnalyses}
\title{Number of factors for a model.}
\arguments{
\item{this}{the FactorAnalyses object.}

\item{model}{the model number.}
}
\description{
Given a model number returns the number of factors in that model
}

