% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getNumLeaves}
\alias{getNumLeaves}
\alias{getNumLeaves.LatentForests}
\title{Get number of leaves.}
\usage{
getNumLeaves(this)

\method{getNumLeaves}{LatentForests}(this)
}
\arguments{
\item{this}{the LatentForests object.}
}
\description{
Gets the number of leaves in the latent forest models.
}

