% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getNumVertices}
\alias{getNumVertices}
\alias{getNumVertices.LatentForests}
\title{Maximum number of vertices.}
\usage{
getNumVertices(this)

\method{getNumVertices}{LatentForests}(this)
}
\arguments{
\item{this}{the LatentForests object.}
}
\description{
A private method for LatentForests that computes the number
of vertices a tree with this$getNumLeaves() number of leaves has.
}

