% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureModels.R
\name{getPhi}
\alias{getPhi}
\alias{getPhi.MixtureModels}
\title{Get the phi parameter.}
\usage{
getPhi(this, phi)

\method{getPhi}{MixtureModels}(this, phi)
}
\arguments{
\item{this}{the MixtureModels object.}

\item{phi}{the new phi value.}
}
\description{
Gets the phi parameter controlling the strength of the sBIC penalty.
}

