% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/ModelPoset.R
\name{getPrior.BinomialMixtures}
\alias{getPrior}
\alias{getPrior.BinomialMixtures}
\title{The prior on the models.}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns the unnormalized prior on the collection.
}

