% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getSamplingCovMat}
\alias{getSamplingCovMat}
\alias{getSamplingCovMat.LatentForests}
\title{Sampling covariance matrix.}
\usage{
getSamplingCovMat(this)

\method{getSamplingCovMat}{LatentForests}(this)
}
\arguments{
\item{this}{the LatentForests object.}
}
\description{
Returns the sampling covariance matrix for the data set with setData().
}

