% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{getSupport}
\alias{getSupport}
\alias{getSupport.LatentForests}
\title{Get support for a given model.}
\usage{
getSupport(this, model)

\method{getSupport}{LatentForests}(this, model)
}
\arguments{
\item{this}{the LatentForests object.}

\item{model}{the model number.}
}
\description{
Given a model number returns the support of the model. Let E by the matrix
of edges returned by this$getAllEdges(), the support is represented by a
0-1 vector v where the ith entry of v is 1 if the ith edge in E is in the
model and is 0 otherwise.
}

