% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{LatentForests}
\alias{LatentForests}
\title{Construct a poset of gaussian latent forest models.}
\usage{
LatentForests(numLeaves = 0, E = matrix(numeric(0), ncol = 2))
}
\arguments{
\item{numLeaves}{the number of observed variables (these are the leaves of
the model)}

\item{E}{a 2xm matrix of edges corresponding to the edges of the 'super
forest' f for which we compute all subforests. f should have nodes
1:numLeaves as leaves of the forest with no internal nodes as leaves.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of gaussian latent forest models
which have a natural ordering upon them. In particular, we consider the
collection of all latent forests that are subforests of an input forest.
}

