% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCAs.R
\name{LCAs}
\alias{LCAs}
\title{Construct a poset of latent class analysis models.}
\usage{
LCAs(maxNumClasses = 1, numVariables = 2,
                        numStatesForVariables = 2, phi = 1)
}
\arguments{
\item{maxNumClasses}{the number of classes in the largest LCA model to
considered.}

\item{numVariables}{the number of observed variables.}

\item{numStatesForVariables}{the number of states for each observed variable,
at the moment these must all be equal.}

\item{phi}{parameter determining the prior placed on latent class
probabilities.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of LCA models which
have a natural ordering upon them.
}

