% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/MixtureModels.R,
%   R/FactorAnalyses.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{learnCoef}
\alias{learnCoef}
\alias{learnCoef.MixtureModels}
\alias{learnCoef.FactorAnalyses}
\alias{learnCoef.LatentForests}
\alias{learnCoef.ReducedRankRegressions}
\title{Learning coefficient}
\usage{
learnCoef(this, superModel, subModel)
}
\arguments{
\item{this}{the object representing the model poset.}

\item{superModel}{the larger model of the two input models.}

\item{subModel}{the submodel of the larger model.}
}
\description{
Computes the learning coefficient for a model with respect to one of the
model's submodels.
}
