% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{logLike}
\alias{logLike}
\alias{logLike.LatentForests}
\title{Multivariate gaussian log-likelihood.}
\usage{
logLike(this, covMat)

\method{logLike}{LatentForests}(this, covMat)
}
\arguments{
\item{this}{the LatentForests object.}

\item{covMat}{a covariance matrix.}
}
\description{
A private method that returns the log-likelihood of the data set with
setData() under a multivariate gaussian model with a given covariance matrix
and assumed 0 means.
}

