% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtures.R, R/ReducedRankRegressions.R
\name{logLikeMleHelper}
\alias{logLikeMleHelper}
\alias{logLikeMleHelper.GaussianMixtures}
\alias{logLikeMleHelper.ReducedRankRegressions}
\title{A helper function.}
\arguments{
\item{this}{the ReducedRankRegressions object.}

\item{model}{the model number.}
}
\description{
A helper function.

A private method that acts as a helper function for the logLikeMLE method
in the ReducedRankRegressions class.
}

