% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureModels.R
\name{MixtureModels}
\alias{MixtureModels}
\title{Linear collections of mixture models.}
\usage{
MixtureModels(phi = "default")
}
\arguments{
\item{phi}{parameter determining the prior placed on latent class
probabilities.}
}
\value{
An object representing the collection.
}
\description{
An abstract class representing a collection of mixture models that are
linearly ordered by the number of components in the mixture. This class
should not be instantiated, just extended.
}

