% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Estimate the precision matrix for multinormal distribution with given adjacency matrix}
\usage{
mle(data, priori)
}
\arguments{
\item{data}{An \code{n} by \code{p} matrix of observations}

\item{priori}{A \code{p} by \code{p} adjacency matrix of the prior network}
}
\value{
A \code{p} by \code{p} matrix of the estimated maximum likelihood precision matrix
}
\description{
Estimate the precision matrix for multinormal distribution with given adjacency matrix
}
\details{
The methods  are based on the relationship between precision matrix of
  multinormal distribution and regression coefficients.
}
\examples{
# set.seed(1)
# d=simulate(p=10,n=100, prob1 = 0.1, prob2 = 0.2,ka=4)
# data=d$data
# priori=d$realnetwork
# precision=mle(data=data,priori=priori)
}
\author{
Jie Zhou
}
