% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateset.R
\name{modelSet}
\alias{modelSet}
\title{Construct the model pool based on a two-step algorithm}
\usage{
modelSet(data, lambda, P)
}
\arguments{
\item{data}{An \code{n x p} data frame}

\item{lambda}{Tuning parameter vector}

\item{P}{Prior adjacency matrix}
}
\value{
\item{webb}{A list of all the candidate models in the model pool. Each model is represented by a \code{p x p}
  adjacency matrix}
}
\description{
For a given prior graph, use the two step algorithm (including edge enrichment)
  and pruning) to construct the model pool
}
\examples{
 # library("glmnet")
 # set.seed(1)
 # d <- simulate(p = 10, n = 100, prob1 = 0.1, prob2 = 0.2, ka = 4)
 # data <- d$data
 # P <- d$priornetwork
 # lambda <- exp(seq(-5,5, length = 50))
 # pool = modelSet(data = data, lambda = lambda, P = P)
}
\author{
Jie Zhou
}
