% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/ModelPoset.R
\name{parents.BinomialMixtures}
\alias{parents}
\alias{parents.BinomialMixtures}
\title{Parents of a model.}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the parents should be found.}
}
\description{
Returns the immediate parents of a given model, i.e. those models
M that are (in the poset ordering) less than the given model but for
which there exists no other model M' such that M < M' < (given model).
}

