% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedRankRegressions.R
\name{ReducedRankRegressions}
\alias{ReducedRankRegressions}
\title{Construct a poset of reduced rank regression models.}
\usage{
ReducedRankRegressions(numResponses, numCovariates, maxRank)
}
\arguments{
\item{numResponses}{the number of response variables.}

\item{numCovariates}{the number of covariates.}

\item{maxRank}{the largest rank model to be included in the collection.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of reduced rank regression models
which have a natural ordering upon them.
}

