% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbic.R
\name{sBIC}
\alias{sBIC}
\title{Compute the sBIC.}
\usage{
sBIC(X, mod)
}
\arguments{
\item{X}{the data for which the maximum likelihood estimates will be computed
for the given collection of models. To see how this data should be
formatted check the documentation for setData.YourModelName.}

\item{mod}{an object representing a poset of models of the same type, e.g.
a collection of binomial mixture models. The currently implemented
models include:
\itemize{
  \item{Binomial mixtures}{see ?BinomialMixtures}
  \item{Latent class analysis}{see ?LCAs}
  \item{Latent gaussian forests}{see LatentForests}
  \item{Reduced rank regression}{see ReducedRankRegressions}
  \item{1-dimensional gaussian mixtures}{see GaussianMixtures}
}}
}
\value{
A named list containing the components
        \itemize{
         \item{logLike}{the computed MLE log-likelihoods for each model.}
         \item{sBIC}{the sBIC score for each model.}
         \item{BIC}{the usual BIC score for each model.}
         \item{modelPoset}{the input model poset mod.}
        }
}
\description{
Computes the sBIC for a given collection of models
}

