% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R
\name{setData.BinomialMixtures}
\alias{setData.BinomialMixtures}
\title{Set data for the binomial mixture models.}
\usage{
\method{setData}{BinomialMixtures}(this, data)
}
\arguments{
\item{this}{the BinomialMixtures object.}

\item{data}{the data to be set, should be a numeric vector of non-negative
integers.}
}
\description{
Sets the data to be used by the binomial mixture models when computing MLEs.
}

