% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtures.R
\name{setData.GaussianMixtures}
\alias{setData.GaussianMixtures}
\title{Set data for the gaussian mixture models.}
\arguments{
\item{this}{the GaussianMixtures object.}

\item{X}{the data to be set, should be a numeric vector of observations.}
}
\description{
Sets the data to be used by the gaussian mixture models when computing MLEs.
}

