% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentForests.R
\name{setData.LatentForests}
\alias{setData.LatentForests}
\title{Set data for the latent forest models.}
\usage{
\method{setData}{LatentForests}(this, data)
}
\arguments{
\item{this}{the LatentForests object.}

\item{data}{the data to be set, should matrix of observed values where each row
corresponds to a single sample.}
}
\description{
Sets the data to be used by the latent forest models models for computing
MLEs.
}

