% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCAs.R
\name{setData.LCAs}
\alias{setData.LCAs}
\title{Set data for the LCA models.}
\usage{
\method{setData}{LCAs}(this, data)
}
\arguments{
\item{this}{the LCAs object.}

\item{data}{the data to be set, should be an integer valued matrix where each
row represents a single sample from the observed variables.}
}
\description{
Sets the data to be used by the LCA models when computing MLEs.
}

