% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedRankRegressions.R
\name{setData.ReducedRankRegressions}
\alias{setData.ReducedRankRegressions}
\title{Set data for the reduced rank regression models.}
\usage{
\method{setData}{ReducedRankRegressions}(this, data)
}
\arguments{
\item{this}{the ReducedRankRegressions object.}

\item{data}{the data to be set, should be a named list with two components:
\itemize{
  \item{X}{A matrix containing the values of covariates for each
          sample. Here each COLUMN represents a single sample from
          all of the covariates.}
  \item{Y}{A matrix containing the values of the response variables
           for each sample. Again, each COLUMN is a single sample.}
}}
}
\description{
Sets the data to be used by the reduced rank regression models when computing
MLEs.
}

