% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCAs.R
\name{setPhi}
\alias{setPhi}
\alias{setPhi.LCAs}
\title{Set phi parameter.}
\arguments{
\item{this}{the LCAs object.}

\item{phi}{the new phi value.}
}
\description{
Set the phi parameter in an LCAs object to a different value.
}

