% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGGM.R
\name{sGGM}
\alias{sGGM}
\title{Model selection of Gaussian graphical model based on SBIC}
\usage{
sGGM(data, lambda, M, prob)
}
\arguments{
\item{data}{A \code{n} by \code{p} data frame representing the observations}

\item{lambda}{The tuning parameter vector}

\item{M}{The prior adjacency matrix}

\item{prob}{The mean error rate}
}
\value{
\item{estimate}{the final selected adjacency matrix}
  \item{web}{the model pool}
}
\description{
Select the model  based on the SBIC and the model pool construted by two-step algorithm
}
\examples{
set.seed(1)
# d=simulate(p=10,n=100, prob1 = 0.1, prob2 = 0.2,ka=4)
# lambda=exp(seq(-5,5,length=50))
# data=d$data
# M=d$priornetwork
# prob=0.2
# final=sGGM(data=data, lambda=lambda, M=M,prob=prob)
# final$estimate
}
