% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSimulate.R
\name{simulate}
\alias{simulate}
\title{Randomly generate adjacent matrices for data simulation purposes}
\usage{
simulate(p, n, prob1, prob2, ka)
}
\arguments{
\item{p}{Number of vertices in the graph}

\item{n}{Sample size}

\item{prob1}{The densitiy of the edge in the prior graph}

\item{prob2}{The discrepancy rate between true and prior edge density}

\item{ka}{The diagonal entry in the true precision matrix}
}
\value{
A list including the simulated data, real adjacency matrix and a prior adjacency matrix.
  \item{n}{sample size}
  \item{prob1}{the density of edge in prior graph}
  \item{prob2}{the discrepancy rate between true and prior edge density}
  \item{ka}{the diagonal entry in real precision matrix}
}
\description{
For a given edge density, first generate an adjacency matrix \code{P} of a graph.
  Based on \code{P}, generate multinomial data with mean 0 and a given precision matrix.
}
\details{
The precision matrix has all off-diagonal entries to be 1 and the diagonal entries to be \code{ka}
}
\examples{
# set.seed(1)
# d=simulate(p=10,n=100,prob1=0.2, prob2=0.1, ka=4)
# d$data
# d$realnetwork
# d$priornetwork
}
\author{
Jie Zhou
}
