% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R
\name{BinomialMixtures}
\alias{BinomialMixtures}
\title{Construct a poset of binomial mixture models.}
\usage{
BinomialMixtures(maxNumComponents = 1, phi = "default")
}
\arguments{
\item{maxNumComponents}{the maximum number of components allowed in a model, will
create a hierarchy of all models with less than or equal
to this number.}

\item{phi}{parameter controlling the strength of the sBIC penalty.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of binomial mixture models. There
is one model for each fixed number of components from 1 to some specified
maximum. In particular each model is identified by a single number
specifiying the number of components in the model. Models are naturally
ordered by inclusion so that, for example, a model with 2 components comes
before a model with 3 or more components.
}

