% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R
\name{BinomialMixtures}
\alias{BinomialMixtures}
\title{Construct a poset of binomial mixture models.}
\usage{
BinomialMixtures(numCovariates = 1, maxNumFactors = 0,
                        phi = "default")
}
\arguments{
\item{numCovariates}{the number of covariates in all of the models.}

\item{maxNumFactors}{the maximum number of factors allowed in a model, will
create a hierarchy of all models with less than or equal
to this number.}

\item{phi}{parameter determining the prior placed on latent class
probabilities.}
}
\value{
An object representing the collection.
}
\description{
Creates an object representing a collection of binomial mixture models. There
is one model for each fixed number of components from 1 to some specified
maximum. In particular each model is identified by a single number
specifiying the number of components in the model. Models are naturally
ordered by inclusion so that, for example, a model with 2 components comes
before a model with 3 or more components.
}

