% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Comparing the real and estimated adjacency matrix}
\usage{
comparison(real, estimate)
}
\arguments{
\item{real}{The real matrix \code{p} by \code{p} adjacency matrix likely from simulated data}

\item{estimate}{The estimated matrix \code{p} by \code{p} adjacency matrix likely estimated using the SBIC procedure}
}
\value{
A list of the following evaluation metrics 
  \item{PSR}{Positive Selection Rate}
  \item{FDR}{False Discovery rate}
}
\description{
Comparing the two adjacency matrices for false discovery rate and positive selection rate. Used
  for model validation
}
\author{
Jie Zhou
}
