% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateset.R
\name{deletion}
\alias{deletion}
\title{Pruning step for constructing the model pool}
\usage{
deletion(data, lambda, P)
}
\arguments{
\item{data}{An \code{n} by \code{p} matrix of observations}

\item{lambda}{Vector of tuning parameter}

\item{P}{Prior adjacency matrix}
}
\value{
A list of model objects
}
\description{
This is the pruning step in the two-step algorithm to construct the model pool (internal use only)
}
\author{
Jie Zhou
}
