% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getData}
\alias{getData}
\alias{getData.BinomialMixtures}
\alias{getData.FactorAnalyses}
\alias{getData.GaussianMixtures}
\alias{getData.LCAs}
\alias{getData.LatentForests}
\alias{getData.ReducedRankRegressions}
\title{Return the set data.}
\usage{
getData(this)

\method{getData}{BinomialMixtures}(this)

\method{getData}{FactorAnalyses}(this)

\method{getData}{GaussianMixtures}(this)

\method{getData}{LCAs}(this)

\method{getData}{LatentForests}(this)

\method{getData}{ReducedRankRegressions}(this)
}
\arguments{
\item{this}{the object from which to get the data.}
}
\description{
If data has been set for the object using the setData() function
then will return that data, otherwise will throw an error.
}

