% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/ModelPoset.R
\name{getData.BinomialMixtures}
\alias{getData}
\alias{getData.BinomialMixtures}
\title{Return the set data.}
\arguments{
\item{this}{the BinomialMixtures object}
}
\description{
If data has been set for the object using the getData() function
then will return that data, otherwise will throw an error.
}

