% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getDimension}
\alias{getDimension}
\alias{getDimension.BinomialMixtures}
\alias{getDimension.FactorAnalyses}
\alias{getDimension.GaussianMixtures}
\alias{getDimension.LCAs}
\alias{getDimension.LatentForests}
\alias{getDimension.ReducedRankRegressions}
\title{Model dimension.}
\usage{
getDimension(this, model)

\method{getDimension}{BinomialMixtures}(this, model)

\method{getDimension}{FactorAnalyses}(this, model)

\method{getDimension}{GaussianMixtures}(this, model)

\method{getDimension}{LCAs}(this, model)

\method{getDimension}{LatentForests}(this, model)

\method{getDimension}{ReducedRankRegressions}(this, model)
}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the dimension should be computed.}
}
\description{
Computes the dimension of a model in the model poset.
}

