% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LatentForests.R, R/ModelPoset.R
\name{getDimension.BinomialMixtures}
\alias{getDimension}
\alias{getDimension.BinomialMixtures}
\alias{getDimension.FactorAnalyses}
\alias{getDimension.GaussianMixtures}
\alias{getDimension.LatentForests}
\title{Model dimension.}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the dimension should be computed.}
}
\description{
Computes the dimension of a model in the model poset.
}

