% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getNumModels}
\alias{getNumModels}
\alias{getNumModels.BinomialMixtures}
\alias{getNumModels.FactorAnalyses}
\alias{getNumModels.GaussianMixtures}
\alias{getNumModels.LCAs}
\alias{getNumModels.LatentForests}
\alias{getNumModels.ReducedRankRegressions}
\title{Number of models.}
\usage{
getNumModels(this)

\method{getNumModels}{BinomialMixtures}(this)

\method{getNumModels}{FactorAnalyses}(this)

\method{getNumModels}{GaussianMixtures}(this)

\method{getNumModels}{LCAs}(this)

\method{getNumModels}{LatentForests}(this)

\method{getNumModels}{ReducedRankRegressions}(this)
}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns the number of models in the collection.
}

