% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LatentForests.R, R/ModelPoset.R
\name{getNumModels.BinomialMixtures}
\alias{getNumModels}
\alias{getNumModels.BinomialMixtures}
\alias{getNumModels.FactorAnalyses}
\alias{getNumModels.GaussianMixtures}
\alias{getNumModels.LatentForests}
\title{Number of models.}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns the number of models in the collection.
}

