% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getNumSamples}
\alias{getNumSamples}
\alias{getNumSamples.BinomialMixtures}
\alias{getNumSamples.FactorAnalyses}
\alias{getNumSamples.GaussianMixtures}
\alias{getNumSamples.LCAs}
\alias{getNumSamples.LatentForests}
\alias{getNumSamples.ReducedRankRegressions}
\title{Number of samples in the set data.}
\usage{
getNumSamples(this)

\method{getNumSamples}{BinomialMixtures}(this)

\method{getNumSamples}{FactorAnalyses}(this)

\method{getNumSamples}{GaussianMixtures}(this)

\method{getNumSamples}{LCAs}(this)

\method{getNumSamples}{LatentForests}(this)

\method{getNumSamples}{ReducedRankRegressions}(this)
}
\arguments{
\item{this}{the object from which to get the number of samples.}
}
\description{
If data has been set using the setData method then returns the
number of samples in the data. Otherwise throws an error.
}

