% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LatentForests.R, R/ModelPoset.R
\name{getNumSamples.BinomialMixtures}
\alias{getNumSamples}
\alias{getNumSamples.BinomialMixtures}
\alias{getNumSamples.FactorAnalyses}
\alias{getNumSamples.GaussianMixtures}
\alias{getNumSamples.LatentForests}
\title{Number of samples in the set data.}
\arguments{
\item{this}{the BinomialMixtures object.}
}
\description{
If data has been set using the setData method then returns the
number of samples in the data. Otherwise throws an error.
}

