% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getPrior}
\alias{getPrior}
\alias{getPrior.BinomialMixtures}
\alias{getPrior.FactorAnalyses}
\alias{getPrior.GaussianMixtures}
\alias{getPrior.LCAs}
\alias{getPrior.LatentForests}
\alias{getPrior.ReducedRankRegressions}
\title{The prior on the models.}
\usage{
getPrior(this)

\method{getPrior}{BinomialMixtures}(this)

\method{getPrior}{FactorAnalyses}(this)

\method{getPrior}{GaussianMixtures}(this)

\method{getPrior}{LCAs}(this)

\method{getPrior}{LatentForests}(this)

\method{getPrior}{ReducedRankRegressions}(this)
}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns the unnormalized prior on the collection.
}

