% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{getTopOrder}
\alias{getTopOrder}
\alias{getTopOrder.BinomialMixtures}
\alias{getTopOrder.FactorAnalyses}
\alias{getTopOrder.GaussianMixtures}
\alias{getTopOrder.LCAs}
\alias{getTopOrder.LatentForests}
\alias{getTopOrder.ReducedRankRegressions}
\title{Topological ordering of models.}
\usage{
getTopOrder(this)

\method{getTopOrder}{BinomialMixtures}(this)

\method{getTopOrder}{FactorAnalyses}(this)

\method{getTopOrder}{GaussianMixtures}(this)

\method{getTopOrder}{LCAs}(this)

\method{getTopOrder}{LatentForests}(this)

\method{getTopOrder}{ReducedRankRegressions}(this)
}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns a topological ordering of models in the collection.
}

