% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LatentForests.R, R/ModelPoset.R
\name{getTopOrder.BinomialMixtures}
\alias{getTopOrder}
\alias{getTopOrder.BinomialMixtures}
\alias{getTopOrder.FactorAnalyses}
\alias{getTopOrder.GaussianMixtures}
\alias{getTopOrder.LatentForests}
\title{Topological ordering of models.}
\arguments{
\item{this}{the model poset object.}
}
\description{
Returns a topological ordering of models in the collection.
}

