% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LatentForests.R, R/ModelPoset.R
\name{learnCoef.BinomialMixtures}
\alias{learnCoef}
\alias{learnCoef.BinomialMixtures}
\alias{learnCoef.FactorAnalyses}
\alias{learnCoef.GaussianMixtures}
\alias{learnCoef.LatentForests}
\title{Learning coefficient}
\arguments{
\item{this}{the object representing the model poset.}

\item{superModel}{the larger model of the two input models.}

\item{subModel}{the submodel of the larger model.}
}
\description{
Computes the learning coefficient for a model with respect to one of the
model's submodels.
}

