% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/ModelPoset.R
\name{logLikeMle.BinomialMixtures}
\alias{logLikeMle}
\alias{logLikeMle.BinomialMixtures}
\title{Maximum likelihood for data.}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the maximum likelihood should be computed.}
}
\description{
Computes the maximum likelihood of a model in the model poset for the
data set using the setData command.
}

