% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedRankRegressions.R
\name{logLikeMleHelper}
\alias{logLikeMleHelper}
\alias{logLikeMleHelper.ReducedRankRegressions}
\title{Help compute the MLE.}
\usage{
logLikeMleHelper(this, model)

\method{logLikeMleHelper}{ReducedRankRegressions}(this, model)
}
\arguments{
\item{this}{the ReducedRankRegressions object.}

\item{model}{the model number.}
}
\description{
A private method that acts as a helper function for the logLikeMLE method
in the ReducedRankRegressions class.
}

