% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialMixtures.R, R/FactorAnalyses.R, R/GaussianMixtures.R, R/LCAs.R, R/LatentForests.R, R/ModelPoset.R
\name{mle.BinomialMixtures}
\alias{mle}
\alias{mle.BinomialMixtures}
\alias{mle.FactorAnalyses}
\alias{mle.GaussianMixtures}
\alias{mle.LCAs}
\alias{mle.LatentForests}
\title{Maximum likelihood estimator.}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the maximum likelihood should be computed.}
}
\description{
Computes the maximum likelihood estimator of the model parameters (for a
given model in the collection) given the data set with setData.
}

