% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R, R/MixtureModels.R, R/FactorAnalyses.R, R/LatentForests.R, R/ReducedRankRegressions.R
\name{parents}
\alias{parents}
\alias{parents.FactorAnalyses}
\alias{parents.LatentForests}
\alias{parents.MixtureModels}
\alias{parents.ReducedRankRegressions}
\title{Parents of a model.}
\usage{
parents(this, model)

\method{parents}{MixtureModels}(this, model)

\method{parents}{FactorAnalyses}(this, model)

\method{parents}{LatentForests}(this, model)

\method{parents}{ReducedRankRegressions}(this, model)
}
\arguments{
\item{this}{the object representing the model poset.}

\item{model}{the model for which the parents should be found.}
}
\description{
Returns the immediate parents of a given model, i.e. those models
M that are (in the poset ordering) less than the given model but for
which there exists no other model M' such that M < M' < (given model).
}

