% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbic.R
\name{sBIC-package}
\alias{sBIC-package}
\title{sBIC package documentation.}
\description{
Computes the sBIC for various model collections including:
\itemize{
 \item{Binomial mixtures}
 \item{Factor analyses}
 \item{Gaussian mixtures}
 \item{Latent forests}
 \item{Latent class analyses}
 \item{Reduced rank regressions}
}
The primary functionality of this package can be accessed through the
sBIC function.
}

