% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtures.R
\name{setData.GaussianMixtures}
\alias{setData.GaussianMixtures}
\title{Set data for the gaussian mixture models.}
\usage{
\method{setData}{GaussianMixtures}(this, data)
}
\arguments{
\item{this}{the GaussianMixtures object.}

\item{data}{the data to be set, a matrix where each row corresponds to a single
multivariate observation. If the corresponding GaussianMixtures object
has ambient dimension 1, then data may be a numeric vector of
observations.}
}
\description{
Sets the data to be used by the gaussian mixture models when computing MLEs.
}

