% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoset.R
\name{setData}
\alias{setData}
\title{Set data for a model poset.}
\usage{
setData(this, data)
}
\arguments{
\item{this}{the model poset object.}

\item{data}{the data to be set.}
}
\description{
Sets the data to be used by a poset of models when computing MLEs.
}

