% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureModels.R
\name{setPhi}
\alias{setPhi}
\alias{setPhi.MixtureModels}
\title{Set phi parameter.}
\usage{
setPhi(this, phi)

\method{setPhi}{MixtureModels}(this, phi)
}
\arguments{
\item{this}{the MixtureModels object.}

\item{phi}{the new phi value.}
}
\description{
Set the phi parameter in a mixture model object to a different value.
}

